Version 5.1.5

+++ Source/UnitFormMain.pas
@@ -390,7 +390,7 @@ var
   wTimeOfDay: Integer;
   wMinutes: Integer;
   wNextSeason: TDateTime;
-  wServerTick: Integer;
+  wServerTick: Int64;
 begin
   wStream := TMemoryStream.Create;
   wXmlDoc := TXpObjModel.Create(nil);
